Filter out features in PD/TMT data

If inputLevel is "Proteins", exclude features with 'Score.Sequest.HT.Sequest.HT' below minScore, 'Number.of.Peptides' below minPeptides, or identified as 'Contaminant' by ProteomeDiscoverer. If inputLevel is "PeptideGroups", exclude features with 'Delta.Score.by.Search.Engine.Sequest.HT' below minDeltaScore, 'Number.of.PSMs' below minPSMs, or identified as 'Contaminant' by ProteomeDiscoverer.

filterPDTMT(
  sce,
  inputLevel,
  minScore = 0,
  minPeptides = 0,
  minDeltaScore = 0,
  minPSMs = 0,
  masterProteinsOnly = FALSE,
  plotUpset = TRUE,
  exclFile = NULL
)

Arguments

sce

A SummarizedExperiment object (or a derivative).

inputLevel

Either "Proteins" or "PeptideGroups", indicating the type of features in sce.

minScore

Numeric scalar, the minimum allowed value in the 'Score.Sequest.HT.Sequest.HT' column in order to retain the feature. Only used if inputLevel is "Proteins".

minPeptides

Numeric scalar, the minimum allowed value in the 'Number.of.Peptides' column in order to retain the feature. Only used if inputLevel is "Proteins".

minDeltaScore

Numeric scalar, the minimum allowed value in the 'Delta.Score.by.Search.Engine.Sequest.HT' column in order to retain the feature. Only used if inputLevel is "PeptideGroups".

minPSMs

Numeric scalar, the minimum allowed value in the 'Number.of.PSMs' column in order to retain the feature. Only used if inputLevel is "PeptideGroups".

masterProteinsOnly

Logical scalar indicating whether only master proteins (where the Master column value is IsMasterProtein) should be retained.

plotUpset

Logical scalar, whether to generate an UpSet plot detailing the reasons for features being filtered out. Only generated if any feature is in fact filtered out.

exclFile

Character scalar, the path to a text file where the features that are filtered out are written. If NULL (default), excluded features are not recorded.

Value

A filtered object of the same type as sce.

Author

Charlotte Soneson