Filter out features in Spectronaut data

Exclude features with 'PG.Cscore' below minScore, 'PG.NrOfStrippedSequencesIdentified.Experiment.wide' below minPeptides, or where the 'PG.ProteinGroups' column contains the specified revPattern.

filterSpectronaut(
  sce,
  minScore,
  minPeptides,
  plotUpset = TRUE,
  revPattern = "_Decoy$",
  contamPattern = "^contam_",
  exclFile = NULL
)

Arguments

sce

A SummarizedExperiment object (or a derivative).

minScore

Numeric scalar, the minimum allowed value in the 'PG.Cscore' column in order to retain the feature.

minPeptides

Numeric scalar, the minimum allowed value in the 'PG.NrOfStrippedSequencesIdentified.Experiment.wide' column in order to retain the feature.

plotUpset

Logical scalar, whether to generate an UpSet plot detailing the reasons for features being filtered out. Only generated if any feature is in fact filtered out.

revPattern

Character scalar providing the pattern (a regular expression) used to identify decoys (reverse hits). The pattern is matched against the IDs in the Spectronaut PG.ProteinGroups column.

contamPattern

Character scalar providing the pattern (a regular expression) used to identify contaminants. The pattern is matched against the IDs in the Spectronaut PG.ProteinGroups column.

exclFile

Character scalar, the path to a text file where the features that are filtered out are written. If NULL (default), excluded features are not recorded.

Value

A filtered object of the same type as sce.

Author

Charlotte Soneson