Format table columns

Format columns in a data.frame.

formatTableColumns(tbl, columns, signifDigits, maxLevels = 10)

Arguments

tbl

A data.frame or similar object.

columns

Character vector giving the column names of tbl that should be formatted.

signifDigits

Numeric scalar, the number of significant digits to round numeric columns to. Set to NULL to skip rounding.

maxLevels

Numeric scalar. If character columns have at most this number of unique values, they will be encoded as factors.

Value

The input data.frame with formatted columns.

Author

Charlotte Soneson

Examples

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y"),
                         signifDigits = 2, maxLevels = 10)
df
#>        x y           z
#> 1  -0.38 A  0.40329035
#> 2  -0.58 A  1.22849759
#> 3  -0.36 C -2.95741334
#> 4  -1.00 A  1.80830093
#> 5   0.13 A -1.58868753
#> 6  -0.59 A  0.97213814
#> 7  -0.79 A  1.61917955
#> 8   0.58 B  0.68085430
#> 9  -0.11 B  0.10106382
#> 10  0.21 A -0.06874891
summary(df)
#>        x           y           z          
#>  Min.   :-1.0000   A:7   Min.   :-2.9574  
#>  1st Qu.:-0.5875   B:2   1st Qu.:-0.0263  
#>  Median :-0.3700   C:1   Median : 0.5421  
#>  Mean   :-0.2890         Mean   : 0.2198  
#>  3rd Qu.: 0.0700         3rd Qu.: 1.1644  
#>  Max.   : 0.5800         Max.   : 1.8083  

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y", "z"),
                         signifDigits = 2, maxLevels = 2)
df
#>         x y      z
#> 1  -0.170 B  1.000
#> 2  -0.610 C  0.270
#> 3   0.078 A  1.400
#> 4   0.012 C  0.011
#> 5   1.100 B  1.800
#> 6   0.300 B  0.310
#> 7  -0.420 B -1.100
#> 8   0.330 C  1.100
#> 9  -0.440 A  1.200
#> 10  1.000 B  0.650
summary(df)
#>        x                y                   z          
#>  Min.   :-0.6100   Length:10          Min.   :-1.1000  
#>  1st Qu.:-0.3575   Class :character   1st Qu.: 0.2800  
#>  Median : 0.0450   Mode  :character   Median : 0.8250  
#>  Mean   : 0.1180                      Mean   : 0.6641  
#>  3rd Qu.: 0.3225                      3rd Qu.: 1.1750  
#>  Max.   : 1.1000                      Max.   : 1.8000