Import an abundance file

Import data from a quantification file (e.g. MaxQuant peptideGroups.txt, Proteome Discoverer Proteins.txt) into a SingleCellExperiment object. Typically sample-specific columns will be used to form assays, and other columns will be added as rowData columns.

importExperiment(
  inFile,
  iColPattern,
  includeOnlySamples = "",
  excludeSamples = "",
  ...
)

Arguments

inFile

Path to a tab-delimited input text file (e.g. MaxQuant peptideGroups.txt, Proteome Discoverer Proteins.txt or FragPipe combined_protein.tsv).

iColPattern

Character scalar defining a regular expression to identify sample columns. For MaxQuant output, this is typically one of "^iBAQ\.", "^LFQ\.intensity\." or "^Intensity\.". For Proteome Discoverer, it is typically "^Abundance\.", "^Abundance\.F[0-9]+\." or "^Abundance\.F.+\.Sample\.". For FragPipe, it is typically "\.MaxLFQ\.Intensity$". Columns matching the given pattern will form the first assay in the output object.

includeOnlySamples, excludeSamples

Character vectors defining regular expressions to match against the extracted columns to retain or exclude samples.

...

Additional arguments that will be passed on to QFeatures::readSummarizedExperiment (e.g., the number of rows to import).

Value

A list with two elements: a SingleCellExperiment object and a character scalar with the main assay name (this assay contains the values from the columns matching the provided iColPattern).

Author

Charlotte Soneson

Examples

sceL <- importExperiment(system.file("extdata", "mq_example",
                                     "1356_proteinGroups.txt",
                                     package = "einprot"),
                         iColPattern = "^iBAQ\\.")
sceL
#> $sce
#> class: SingleCellExperiment 
#> dim: 463 9 
#> metadata(1): colList
#> assays(9): iBAQ MS.MS.Count ... Peptides Identification.type
#> rownames(463): 1 2 ... 462 463
#> rowData names(35): Protein.IDs Majority.protein.IDs ...
#>   Oxidation.M.site.IDs Oxidation.M.site.positions
#> colnames(9): Adnp_IP04 Adnp_IP05 ... RBC_ctrl_IP02 RBC_ctrl_IP03
#> colData names(0):
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
#> 
#> $aName
#> [1] "iBAQ"
#>