Generate a comprehensive database of complexes

This function generates a comprehensive cross-species database of protein complexes. It downloads the complex definitions from http://wodaklab.org/cyc2008/resources/CYC2008_complex.tab (S. cerevisiae), https://mips.helmholtz-muenchen.de/corum/download/releases/current/allComplexes.txt.zip (mammals) and https://www.pombase.org/data/annotations/Gene_ontology/GO_complexes/Complex_annotation.tsv (S. pombe), and next uses the babelgene package to map the complexes to orthologs in the other species. A pre-generated version of the database is provided with einprot (see listComplexDBs()).

makeComplexDB(
  dbDir,
  customComplexTxt = NULL,
  Cyc2008Db = NULL,
  CorumDb = NULL,
  PombaseDb = NULL
)

Arguments

dbDir

Path to database directory, where all raw files will be downloaded and the output will be saved. Will be created if it doesn't exist.

customComplexTxt

File path to text file with custom complexes (if any). Should be a tab-delimited text file with five columns: "Complex.name", "Gene.names", "Organism", "Source", "PMID".

Cyc2008Db, CorumDb, PombaseDb

data.frames providing annotations from CYC2008 (S cerevisiae), Corum (mammals) and Pombase (S pombe), respectively. These arguments are provided mainly to allow testing, and typically are not specified by the end user, except in cases where the files have already been downloaded and stored locally. If provided, it is important that the data frames are obtained by simply reading the downloaded text files - the function assumes a certain set of columns. If NULL (the default), the files will be downloaded from the paths indicated in the Details.

Value

Invisibly, the path to the generated complex database.

Author

Charlotte Soneson

Examples

## Read small subsets of the raw files provided with einprot to make the
## processing faster. Typically, the files would be downloaded as part of
## the process of generating the complex DB.
cyc2008db <- read.delim(system.file("extdata", "complexes",
                                    "cyc2008_complex_extract.tab",
                                    package = "einprot"))
corumdb <- read.delim(system.file("extdata", "complexes",
                                  "corum_complex_extract.txt",
                                  package = "einprot"))
pombasedb <- read.delim(system.file("extdata", "complexes",
                                    "pombase_complex_extract.tsv",
                                    package = "einprot"))
dbdir <- tempdir()
dbs <- makeComplexDB(dbDir = dbdir, Cyc2008Db = cyc2008db,
                     CorumDb = corumdb, PombaseDb = pombasedb)
#> mouse
#> human
#> baker's yeast
#> Caenorhabditis elegans
#> Schizosaccharomyces pombe 972h-

## List of complexes
compl <- readRDS(dbs$complPath)
compl
#> CharacterList of length 10
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] ADA2 ... NGG1
#> [["S.cer: TRAPP complex"]] BET3 BET5 GSG1 KRE11 ... TRS20 TRS23 TRS31 TRS33
#> [["S.cer: histone deacetylase complex"]] HDA1 HDA2 HDA3
#> [["human: BCL6-HDAC4 complex"]] BCL6 HDAC4
#> [["human: BCL6-HDAC5 complex"]] BCL6 HDAC5
#> [["mouse: BLOC-2 (biogenesis of lysosome-related organelles complex 2)"]] Hps...
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] Dnm1l Clta Bcl2l1 Mff
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] swi10 rhp14 rad16
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] rhp41 rhp42
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] tfb1 ptr8 ... pmh1

## Complexes mapped to all species via orthologs
orth <- readRDS(dbs$orthPath)
orth
#> $mouse
#> CharacterList of length 10
#> [["mouse: BLOC-2 (biogenesis of lysosome-related organelles complex 2)"]] Hps...
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] Bcl2l1 Clta Dnm1l Mff
#> [["human: BCL6-HDAC4 complex"]] Bcl6 Hdac4
#> [["human: BCL6-HDAC5 complex"]] Bcl6 Hdac5
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] Kat2b Tada2a
#> [["S.cer: TRAPP complex"]] Trappc1 Trappc2 Trappc3 Trappc4 Trappc5 Trappc6b
#> [["S.cer: histone deacetylase complex"]] Hdac6
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] Ercc1 Ercc4 Xpa
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] Xpc
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] Ercc2 ... Mnat1
#> 
#> $human
#> CharacterList of length 10
#> [["human: BCL6-HDAC4 complex"]] BCL6 HDAC4
#> [["human: BCL6-HDAC5 complex"]] BCL6 HDAC5
#> [["mouse: BLOC-2 (biogenesis of lysosome-related organelles complex 2)"]] HPS...
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] BCL2L1 CLTA DNM1L MFF
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] KAT2B TADA2A
#> [["S.cer: TRAPP complex"]] TRAPPC1 TRAPPC2 TRAPPC3 TRAPPC4 TRAPPC5 TRAPPC6B
#> [["S.cer: histone deacetylase complex"]] HDAC6
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] ERCC1 ERCC4 XPA
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] XPC
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] ERCC2 ... MNAT1
#> 
#> $`baker's yeast`
#> CharacterList of length 7
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] ADA2 ... NGG1
#> [["S.cer: TRAPP complex"]] BET3 BET5 GSG1 KRE11 ... TRS20 TRS23 TRS31 TRS33
#> [["S.cer: histone deacetylase complex"]] HDA1 HDA2 HDA3
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] RAD10 RAD1 RAD14
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] RAD4
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] RAD3 SSL2 ... TFB3
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] DNM1
#> 
#> $`Caenorhabditis elegans`
#> CharacterList of length 7
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] clic-1 drp-1 mff-2
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] pcaf-1
#> [["S.cer: TRAPP complex"]] trpp-1 sedl-1 trpp-3 trpp-4 trpp-5 trpp-6
#> [["S.cer: histone deacetylase complex"]] hda-6
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] ercc-1 xpf-1 xpa-1
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] xpc-1
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] xpd-1 ... mnat-1
#> 
#> $`Schizosaccharomyces pombe 972h-`
#> CharacterList of length 7
#> [["S.pombe: nucleotide-excision repair factor 1 complex"]] swi10 rhp14 rad16
#> [["S.pombe: nucleotide-excision repair factor 2 complex"]] rhp41 rhp42
#> [["S.pombe: nucleotide-excision repair factor 3 complex"]] tfb1 ptr8 ... pmh1
#> [["S.cer: Ada2p/Gcn5p/Ada3 transcription activator complex"]] gcn5 ada2
#> [["S.cer: TRAPP complex"]] bet5 trs20 bet3 trs23 trs31 trs33
#> [["S.cer: histone deacetylase complex"]] clr3
#> [["rat: Bcl2l1-Dnm1l-Mff-Clta complex"]] clc1 dnm1
#>