Make table from list

Construct a nice-looking table from a named list, for inclusion in HTML reports.

makeTableFromList(l)

Arguments

l

Named list. Each element of the list must be a scalar value.

Value

A kableExtra table.

Author

Charlotte Soneson

Examples

## Small list
makeTableFromList(list(first = "one", second = "two", third = 3))
#> <table class=" lightable-paper lightable-striped" style='font-family: "Trebuchet MS", verdana, sans-serif; margin-left: auto; margin-right: auto;'>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> first </td>
#>    <td style="text-align:left;width: 100em; "> one </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> second </td>
#>    <td style="text-align:left;width: 100em; "> two </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> third </td>
#>    <td style="text-align:left;width: 100em; "> 3 </td>
#>   </tr>
#> </tbody>
#> </table>

## MaxQuant parameter table
mq <- readMaxQuantXML(system.file("extdata", "mq_example",
                                  "1356_mqpar.xml",
                                  package = "einprot"))
makeTableFromList(mq)
#> <table class=" lightable-paper lightable-striped" style='font-family: "Trebuchet MS", verdana, sans-serif; margin-left: auto; margin-right: auto;'>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> MaxQuant version </td>
#>    <td style="text-align:left;width: 100em; "> 1.5.3.8 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Parameter file </td>
#>    <td style="text-align:left;width: 100em; "> /Users/runner/work/_temp/Library/einprot/extdata/mq_example/1356_mqpar.xml </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Search engine </td>
#>    <td style="text-align:left;width: 100em; "> Andromeda </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Raw file location </td>
#>    <td style="text-align:left;width: 100em; "> D:/Data/MaxQuant/ </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Raw files </td>
#>    <td style="text-align:left;width: 100em; "> F_160817_AdnpFB_IP04.raw, F_160817_AdnpFB_IP05.raw, F_160817_AdnpFB_IP06.raw, F_160817_Chd4BF_IP07.raw, F_160817_Chd4BF_IP08.raw, F_160817_Chd4BF_IP09.raw, F_160817_RBC_ctrl_IP01.raw, F_160817_RBC_ctrl_IP02.raw, F_160817_RBC_ctrl_IP03.raw </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Sample names </td>
#>    <td style="text-align:left;width: 100em; "> Adnp_IP04, Adnp_IP05, Adnp_IP06, Chd4BF_IP07, Chd4BF_IP08, Chd4BF_IP09, RBC_ctrl_IP01, RBC_ctrl_IP02, RBC_ctrl_IP03 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Databases </td>
#>    <td style="text-align:left;width: 100em; "> D:/Databases/MOUSE__150129.fasta </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Contaminants </td>
#>    <td style="text-align:left;width: 100em; "> */MaxQuant_1.5.3.8/MaxQuant/bin/conf/contaminants.fasta </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Quantification settings (LFQ) </td>
#>    <td style="text-align:left;width: 100em; "> LFQ min. ratio count: 1, fastLFQ: false, match-between runs (MBR): true, Intensity based absolute quantification (iBAQ): true </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Min. razor peptides </td>
#>    <td style="text-align:left;width: 100em; "> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Enzymes </td>
#>    <td style="text-align:left;width: 100em; "> Trypsin/P </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Variable modifications </td>
#>    <td style="text-align:left;width: 100em; "> Oxidation (M), Acetyl (Protein N-term) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Fixed modifications </td>
#>    <td style="text-align:left;width: 100em; ">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Max peptide mass </td>
#>    <td style="text-align:left;width: 100em; "> 8000 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;width: 25em; font-weight: bold;border-right:1px solid;"> Min peptide length </td>
#>    <td style="text-align:left;width: 100em; "> 7 </td>
#>   </tr>
#> </tbody>
#> </table>