Prepare feature collections for testing with camera

Prepare feature collections for testing with camera. The function maps the feature IDs in the collections (complexes or GO terms) to the values in the specified idCol column of rowData(sce), and subsequently replaces them with the corresponding row names of the SummarizedExperiment object. Feature sets with too few features (after the matching) are removed.

prepareFeatureCollections(
  sce,
  idCol,
  includeFeatureCollections,
  complexDbPath,
  speciesInfo,
  complexSpecies,
  customComplexes = list(),
  minSizeToKeep = 2
)

Arguments

sce

A SummarizedExperiment object (or a derivative).

idCol

Character scalar, indicating which column in rowData(sce) that contains IDs matching those in the feature collections (gene symbols).

includeFeatureCollections

Character vector indicating the types of feature collections to prepare. Should be a subset of c("complexes", "GO") or NULL.

complexDbPath

Character scalar providing the path to the database of complexes, generated using makeComplexDB() and serialized to a .rds file.

speciesInfo

List with at least two entries (species and speciesCommon), providing the species information. Typically generated using getSpeciesInfo().

complexSpecies

Character scalar, either "all" or "current", indicating whether all complexes should be tested, or only those defined for the current species.

customComplexes

Named list, for providing any custom complexes that are not already included in the database provided via complexDbPath.

minSizeToKeep

Numeric scalar, indicating the minimum size of a feature collection to be retained.

Value

A list of CharacterLists (one for each feature collection).

Author

Charlotte Soneson