Extract information from DIA-NN logs

Read DIA-NN log file and extract information about the run.

readDIANNInfo(diannLog)

Arguments

diannLog

Character scalar, the path to a log file from a DIA-NN run.

Value

A list with extracted information about the DIA-NN run.

Author

Charlotte Soneson

Examples

readDIANNInfo(system.file("extdata", "diann_example",
                          "diann-output.log.txt",
                          package = "einprot"))
#> $`DIA-NN version`
#> [1] "DIA-NN 1.8.2 beta 8 "
#> 
#> $`DIA-NN log file`
#> [1] "/Users/runner/work/_temp/Library/einprot/extdata/diann_example/diann-output.log.txt"
#> 
#> $`MS2 Mass accuracy (ppm)`
#> [1] "15 "
#> 
#> $`MS1 Mass accuracy (ppm)`
#> [1] "15 "
#> 
#> $`Input files`
#> [1] "/scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_03.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_04.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_06.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_05.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_07.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_08.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_09.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_01.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_02.mzML, /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_03.mzML"
#> 
#> $`Spectral library`
#> [1] "/scratch/cpanse/PXD028735/fasta/uniprotkb_proteome_UP000000625_UP000002311_UP000005640_iRTkit_2023_07_04.2and3plus.predicted.speclib"
#> 
#> $`Sequence databases`
#> [1] "uniprotkb_proteome_UP000000625_2023_07_04.fasta; uniprotkb_proteome_UP000005640_2023_07_04.fasta; uniprotkb_proteome_UP000002311_2023_07_04.fasta; iRTkit.fasta"
#> 
#> $`DIA-NN command`
#> [1] "/usr/diann/1.8.2_beta_8/linux/diann-1.8.1.8 --threads 64 --mass-acc 15 --mass-acc-ms1 15 --matrices --pg-level 1 --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Beta_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Beta_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Alpha_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Alpha_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_A_Sample_Gamma_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Condition_B_Sample_Gamma_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Ecoli_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Human_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_03.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_04.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_06.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_05.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_07.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_08.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_QC_09.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_01.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_02.mzML --f /scratch/cpanse/PXD028735/dia/LFQ_Orbitrap_AIF_Yeast_03.mzML --lib /scratch/cpanse/PXD028735/fasta/uniprotkb_proteome_UP000000625_UP000002311_UP000005640_iRTkit_2023_07_04.2and3plus.predicted.speclib --temp /scratch/cpanse/PXD028735/dia/temp-2023-07-17/ --out /scratch/cpanse/PXD028735/dia/out-2023-07-17//diann-output.tsv "
#>