Extract information from Proteome Discoverer logs

Read metadata from Proteome Discoverer log files. If some of the files are missing, returns the information from the ones that are available.

readProteomeDiscovererInfo(pdOutputFolder, pdResultName, pdAnalysisFile)

Arguments

pdOutputFolder

Character string pointing to the PD/TMT output folder. Should contain the files pdResultName_InputFiles.txt and pdResultName_StudyInformation.txt.

pdResultName

Character string providing the base name for the files in the pdOutputFolder.

pdAnalysisFile

Path to a .pdAnalysis file from Proteome Discoverer. Can be NULL.

Value

A list with settings read from the Proteome Discoverer files.

See also

Author

Charlotte Soneson

Examples

readProteomeDiscovererInfo(
    pdOutputFolder = system.file("extdata", "pdtmt_example",
                                 package = "einprot"),
    pdResultName = "Fig2_m23139_RTS_QC_varMods",
    pdAnalysisFile = system.file("extdata", "pdtmt_example",
                                 "Fig2_m23139_RTS_QC_varMods.pdAnalysis",
                                 package = "einprot"))
#> $`PD version`
#> [1] "2.5.0.400"
#> 
#> $`PD output folder`
#> [1] "/Users/runner/work/_temp/Library/einprot/extdata/pdtmt_example"
#> 
#> $`PD result name`
#> [1] "Fig2_m23139_RTS_QC_varMods"
#> 
#> $`PD analysis file`
#> [1] "/Users/runner/work/_temp/Library/einprot/extdata/pdtmt_example/Fig2_m23139_RTS_QC_varMods.pdAnalysis"
#> 
#> $`PD Processing WF`
#> [1] "PWF_Tribrid_TMTpro_Quan_SPS_MS3_SequestHT_Percolator"
#> 
#> $`PD Consensus WF`
#> [1] "CWF_Comprehensive_Enhanced Annotation_Reporter_Quan"
#> 
#> $`Search engine`
#> [1] "Sequest HT"
#> 
#> $Instruments
#> [1] "Orbitrap Fusion"
#> 
#> $`Raw file location`
#> [1] "D:/Data/PXD017803/"
#> 
#> $`Raw files`
#> [1] "Fig2_m23139_RTS.raw"
#> 
#> $`Sample names`
#> [1] "HIS4KO_S05, HIS4KO_S06, HIS4KO_S07, HIS4KO_S08, MET6KO_S01, MET6KO_S02, MET6KO_S03, MET6KO_S04, URA2KO_S09, URA2KO_S10, URA2KO_S11, URA2KO_S12, WT_S13, WT_S14, WT_S15, WT_S16"
#> 
#> $Databases
#> [1] "CON_iRT_contaminants_cRAPMaxQFMI_150507.fasta; YEAST__210503.fasta"
#> 
#> $Contaminants
#> [1] "CON_iRT_contaminants_cRAPMaxQFMI_150507.fasta"
#> 
#> $`Quantification settings (LFQ)`
#> [1] "Peptides used:Unique + Razor, quan. method: Reporter Ions Quantifier, quan. MS order: MS3, abundance type: S/N, quan. correction: False, MS1 co-isolation threshold: 50, av. reporter SN threshold: 10, PSP mass matches [%]: 65, norm. method: Total Peptide Amount, PD imputation: None"
#> 
#> $Enzymes
#> [1] "Trypsin (Full)"
#> 
#> $`Variable modifications`
#> [1] "Oxidation / +15.995 Da (M), Carbamidomethyl / +57.021 Da (C), TMTpro / +304.207 Da (K, S, T), TMTpro / +304.207 Da (N-Terminus)"
#> 
#> $`Fixed modifications`
#> [1] ""
#> 
#> $`Validation method`
#> [1] "PercolatorConfidenceAssignment"
#> 
#> $`Validation based on`
#> [1] "Target/Decoy, q-Value"
#> 
#> $`Confidence thresholds`
#> [1] "strict: 0.01, relaxed: 0.05"
#> 
#> $`Max missed cleavages`
#> [1] "2"
#>