Run statistical test

Perform pairwise comparisons. If colData(sce) contains a column named 'batch', this will be included as a covariate if testType is 'limma' or 'proDA'. If it contains a column named 'sampleweights', these will be used as sample weights if testType is 'limma'.

runTest(
  sce,
  comparisons,
  groupComposition = NULL,
  testType,
  assayForTests,
  assayImputation = NULL,
  minNbrValidValues = 2,
  minlFC = 0,
  featureCollections = list(),
  complexFDRThr = 0.1,
  volcanoAdjPvalThr = 0.05,
  volcanoLog2FCThr = 1,
  baseFileName = NULL,
  seed = 123,
  samSignificance = TRUE,
  nperm = 250,
  volcanoS0 = 0.1,
  addAbundanceValues = FALSE,
  aName = NULL,
  singleFit = TRUE,
  subtractBaseline = FALSE,
  baselineGroup = "",
  extraColumns = NULL
)

Arguments

sce

A SummarizedExperiment object (or a derivative).

comparisons

A list of character vectors of length 2, each giving the two groups to be compared.

groupComposition

A list providing the composition of each group used in any of the comparisons. If NULL, assumes that each group used in comparisons consists of a single group in the group column of colData(sce).

testType

Character scalar, either "limma", "ttest" or "proDA".

assayForTests

Character scalar, the name of an assay of the SummarizedExperiment object with values that will be used to perform the test.

assayImputation

Character scalar, the name of an assay of sce with logical values indicating whether an entry was imputed or not.

minNbrValidValues

Numeric scalar, the minimum number of valid (non-imputed) values that must be present for a features to include it in the result table.

minlFC

Non-negative numeric scalar, the logFC threshold to use for limma-treat. If minlFC = 0, limma::eBayes is used instead.

featureCollections

List of CharacterLists with feature collections.

complexFDRThr

Numeric scalar giving the significance (FDR) threshold below which a complex will be considered significant.

volcanoAdjPvalThr

Numeric scalar giving the FDR threshold for significance (for later use in volcano plots).

volcanoLog2FCThr

Numeric scalar giving the logFC threshold for significance (for later use in volcano plots).

baseFileName

Character scalar or NULL, the base file name of the output text files. If NULL, no result files are generated.

seed

Numeric scalar, the random seed to use for permutation (only used if testType is "ttest").

samSignificance

Logical scalar, indicating whether the SAM statistic should be used to determine significance (similar to the approach used by Perseus). Only used if testType = "ttest". If FALSE, the p-values are adjusted using the Benjamini-Hochberg approach and used to determine significance.

nperm

Numeric scalar, the number of permutations (only used if testType is "ttest").

volcanoS0

Numeric scalar, the S0 value to use for creating significance curves (only used if testType is "ttest").

addAbundanceValues

Logical scalar, whether to extract abundance and add to the result table.

aName

Character scalar, the name of the assay in the SummarizedExperiment object to get abundance values from (only required if addAbundanceValues is TRUE).

singleFit

Logical scalar, whether to fit a single model to the full data set and extract relevant results using contrasts. If FALSE, the data set will be subset for each comparison to only the relevant samples. Setting singleFit to TRUE is only supported for testType = "limma" or "proDA".

subtractBaseline

Logical scalar, whether to subtract the background/ reference value for each feature in each batch before fitting the model. If TRUE, requires that a 'batch' column is available.

baselineGroup

Character scalar representing the reference group. Only used if subtractBaseline is TRUE, in which case the abundance values for a given sample will be adjusted by subtracting the average value across all samples in the baselineGroup from the same batch as the original sample.

extraColumns

Character vector (or NULL) indicating columns of rowData(sce) to include in the result table.

Value

A list with seven components: tests (a list with test results), plotnotes (the prior df used by limma), plottitles

(indicating the type of test), plotsubtitles (indicating the significance thresholds), featureCollections (list of feature sets, expanded with results from camera), topsets

(a list with the significant feature sets), messages (any messages for the user), design (information about the experimental design) and curveparams (information required to create Perseus-like significance curves). In addition, if baseFileName is not NULL, text files with test results (including only features and feature sets passing the imposed significance thresholds) are saved.

Author

Charlotte Soneson