Generate a shiny app to explore sequence logos

Launch a shiny app that displays the provided csv file as an interactive table, and generates a sequence logo for the sequences in the seqWindow column.

seqLogoApp(
  seqTableCsv,
  exportName = sub("\\.csv$", "", basename(seqTableCsv)),
  ...
)

Arguments

seqTableCsv

Character scalar, the path to a csv file. Typically this will be exported by the PTM einprot workflow. The file must have a column named seqWindow, containing sequences of a fixed width.

exportName

Character scalar defining the default file name of the files exported from the shiny app.

...

Additional arguments passed to utils::read.csv.

Value

A shiny app object.

Author

Charlotte Soneson

Examples

tbl <- data.frame(id = c("seq1", "seq2", "seq3"),
                  seqWindow = c("LITKDHE", "RQMKQPE", "LSNKVHG"))
tfile <- tempfile(fileext = "_seq.csv")
write.table(tbl, file = tfile, row.names = FALSE,
            col.names = TRUE, quote = TRUE, sep = ",")
if (interactive()) {
    seqLogoApp(tfile)
}