Text snippets for use in analysis reports

These text snippets are pasted in the analysis reports and describe the analysis that was performed. The functions are not intended to be called directly by the user.

emptySampleText(sce, assayName)

testText(testType, minlFC = 0, samSignificance = TRUE)

normText(normMethod)

saText(testType)

expDesignText(testType)

introText(expType)

filterByModText(excludeUnmodifiedPeptides, keepModifications)

inputText(expTypeLevel)

featureCollectionText(featureCollections)

Arguments

sce

A SummarizedExperiment object.

assayName

Character scalar representing the assay that will be used to check whether there are samples with missing (NA) values for all features.

testType

Character scalar giving the statistical test, either "limma", "ttest", or "proDA".

minlFC

Numeric scalar giving the minimum logFC threshold.

samSignificance

Logical scalar indicating whether the SAM statistic should be used to determine significance for the t-test.

normMethod

Character scalar giving the normalization method.

expType

The quantification/identification tool used to generate the data, one of "MaxQuant", "ProteomeDiscoverer", "FragPipe", "DIANN" or "Spectronaut".

excludeUnmodifiedPeptides

Logical scalar indicating whether to exclude (filter out) unmodified peptides.

keepModifications

Character string (or NULL) indicating which modifications to retain in the analysis. Can be a regular expression. If NULL (the default), all rows are retained.

expTypeLevel

The quantification/identification tool used to generate the data (including the data level for ProteomeDiscoverer), one of "MaxQuant", "ProteomeDiscovererProteins", "ProteomeDiscovererPeptideGroups", "FragPipe", "DIANN" or "Spectronaut".

featureCollections

Character vector, a subset of c("complexes", "GO", "pathways").

Value

A character string.

Author

Charlotte Soneson

Examples

testText(testType = "limma")
#> [1] "For each feature, we then compare the (possibly imputed) log2 intensities between groups. For this, we use the [limma](https://bioconductor.org/packages/limma/) R/Bioconductor package [@Ritchie2015limma; @Phipson2016robust]. For more information about the df.prior, representing the amount of extra information that is borrowed from the full set of features in order to improve the inference for each feature, see section 13.2 in the [limma user guide](https://www.bioconductor.org/packages/devel/bioc/vignettes/limma/inst/doc/usersguide.pdf). If requested, in addition to the feature-wise tests, we apply the camera method [@Wu2012camera] to test for significance of each included feature collection. These tests are based on the t-statistics returned from limma."
testText(testType = "ttest")
#> [1] "For each feature, we then compare the (possibly imputed) log2 intensities between groups. For this, we use a Student's t-test. To determine which features show significant changes, we calculate the SAM statistic [@Tusher2001sam], and estimate the false discovery rate at different thresholds using permutations, mimicking the approach used by Perseus [@Tyanova2016perseus]. If requested, in addition to the feature-wise tests, we apply the camera method [@Wu2012camera] to test for significance of each included feature collection. These tests are based on the SAM statistics calculated from the t-statistics and the specified S0."
normText(normMethod = "none")
#> [1] "The log2 intensities are not normalized further across samples since 'normMethod' is set to 'none'."
expDesignText(testType = "limma")
#> [1] "The plots below illustrate the experimental design used for the linear model(s) and contrasts by `limma`. The plot to the right shows the number of samples for each combination of factor levels across the predictors, and is useful for detecting imbalances between group sizes for different conditions. The plot to the left summarizes the expected response value for each combination of predictor levels, expressed in terms of the linear model coefficients. For more details on how to interpret the plots, we refer to @Soneson2020emm or @Law2020design. Clicking on the arrow below the plots will reveal the design matrix used by limma, as well as the contrasts that were fit for each comparison."