Format table columns

Format columns in a data.frame.

formatTableColumns(tbl, columns, signifDigits, maxLevels = 10)

Arguments

tbl

A data.frame or similar object.

columns

Character vector giving the column names of tbl that should be formatted.

signifDigits

Numeric scalar, the number of significant digits to round numeric columns to. Set to NULL to skip rounding.

maxLevels

Numeric scalar. If character columns have at most this number of unique values, they will be encoded as factors.

Value

The input data.frame with formatted columns.

Author

Charlotte Soneson

Examples

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y"),
                         signifDigits = 2, maxLevels = 10)
df
#>        x y           z
#> 1   0.50 C -0.03982974
#> 2   0.15 B  1.47774639
#> 3   0.30 C -1.24440612
#> 4  -0.12 B -0.76752290
#> 5  -0.53 B -0.41004913
#> 6   0.21 C -0.26753079
#> 7   0.43 A  0.94822392
#> 8   0.28 B -0.41930054
#> 9   0.64 A -0.44601398
#> 10 -1.70 B -0.54395016
summary(df)
#>        x           y           z           
#>  Min.   :-1.7000   A:2   Min.   :-1.24441  
#>  1st Qu.:-0.0525   B:5   1st Qu.:-0.51947  
#>  Median : 0.2450   C:3   Median :-0.41467  
#>  Mean   : 0.0160         Mean   :-0.17126  
#>  3rd Qu.: 0.3975         3rd Qu.:-0.09675  
#>  Max.   : 0.6400         Max.   : 1.47775  

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y", "z"),
                         signifDigits = 2, maxLevels = 2)
df
#>        x y       z
#> 1   0.40 A  0.3500
#> 2  -0.93 B -0.3900
#> 3   1.90 A -0.4700
#> 4  -0.87 C -1.2000
#> 5   2.00 C -0.0058
#> 6  -0.64 A -0.8400
#> 7  -1.30 B  2.1000
#> 8  -0.87 A -0.4700
#> 9  -0.73 A  0.8500
#> 10  0.61 A -0.1300
summary(df)
#>        x                y                   z           
#>  Min.   :-1.3000   Length:10          Min.   :-1.20000  
#>  1st Qu.:-0.8700   Class :character   1st Qu.:-0.47000  
#>  Median :-0.6850   Mode  :character   Median :-0.26000  
#>  Mean   :-0.0430                      Mean   :-0.02058  
#>  3rd Qu.: 0.5575                      3rd Qu.: 0.26105  
#>  Max.   : 2.0000                      Max.   : 2.10000