Format table columns

Format columns in a data.frame.

formatTableColumns(tbl, columns, signifDigits, maxLevels = 10)

Arguments

tbl

A data.frame or similar object.

columns

Character vector giving the column names of tbl that should be formatted.

signifDigits

Numeric scalar, the number of significant digits to round numeric columns to. Set to NULL to skip rounding.

maxLevels

Numeric scalar. If character columns have at most this number of unique values, they will be encoded as factors.

Value

The input data.frame with formatted columns.

Author

Charlotte Soneson

Examples

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y"),
                         signifDigits = 2, maxLevels = 10)
df
#>         x y           z
#> 1  -0.860 A  0.08639645
#> 2   0.560 A -1.49402409
#> 3   1.200 B -0.07791682
#> 4   0.089 B -0.09334940
#> 5  -0.550 C -0.27210212
#> 6  -0.810 C  1.20571005
#> 7   0.160 B -1.55006679
#> 8   1.800 C -2.14695048
#> 9  -1.500 A  0.27825929
#> 10 -1.800 A  0.26291066
summary(df)
#>        x           y           z           
#>  Min.   :-1.8000   A:4   Min.   :-2.14695  
#>  1st Qu.:-0.8475   B:3   1st Qu.:-1.18854  
#>  Median :-0.2305   C:3   Median :-0.08563  
#>  Mean   :-0.1711         Mean   :-0.38011  
#>  3rd Qu.: 0.4600         3rd Qu.: 0.21878  
#>  Max.   : 1.8000         Max.   : 1.20571  

df <- formatTableColumns(data.frame(x = rnorm(10),
                                    y = sample(LETTERS[seq_len(3)], 10,
                                               replace = TRUE),
                                    z = rnorm(10)),
                         columns = c("x", "y", "z"),
                         signifDigits = 2, maxLevels = 2)
df
#>         x y      z
#> 1  -0.720 A  2.200
#> 2   1.300 A  0.540
#> 3  -0.150 A -0.220
#> 4   0.054 C  0.053
#> 5   0.910 B -0.190
#> 6  -0.160 A -0.021
#> 7   0.700 A -1.900
#> 8  -0.290 B  1.100
#> 9   0.500 C -0.220
#> 10  0.930 A  0.093
summary(df)
#>        x                y                   z          
#>  Min.   :-0.7200   Length:10          Min.   :-1.9000  
#>  1st Qu.:-0.1575   Class :character   1st Qu.:-0.2125  
#>  Median : 0.2770   Mode  :character   Median : 0.0160  
#>  Mean   : 0.3074                      Mean   : 0.1435  
#>  3rd Qu.: 0.8575                      3rd Qu.: 0.4283  
#>  Max.   : 1.3000                      Max.   : 2.2000