Import an abundance file

Import an abundance file

importExperiment(
  inFile,
  iColPattern,
  includeOnlySamples = "",
  excludeSamples = "",
  ...
)

Arguments

inFile

The path to an input text file (e.g. MaxQuant peptideGroups.txt, PD Proteins.txt or FragPipe combined_protein.tsv).

iColPattern

Character scalar defining a regular expression to identify sample columns. For MaxQuant output, this is typically one of "^iBAQ\.", "^LFQ\.intensity\." or "^Intensity\.". For PD, it is typically "^Abundance\.", "^Abundance\.F[0-9]+\." or "^Abundance\.F.+\.Sample\.". For FragPipe, it is typically "\.MaxLFQ\.Intensity$". Columns matching the given pattern will form the first assay in the output object.

includeOnlySamples, excludeSamples

Character vectors defining regular expressions to match against the extracted columns to retain or exclude samples.

...

Additional arguments that will be passed on to QFeatures::readSummarizedExperiment (e.g., the number of rows to import).

Value

A list with two elements: a SingleCellExperiment object and a character scalar with the main assay name (this assay contains the values from the columns matching the provided iColPattern).

Author

Charlotte Soneson

Examples

sce <- importExperiment(system.file("extdata", "mq_example",
                                    "1356_proteinGroups.txt",
                                    package = "einprot"),
                        iColPattern = "^iBAQ\\.")